serial-bridge/requirements.txt
pyserial==3.5
requests==2.31.0
python-dotenv==1.0.0
python-socketio[client]==5.11.0
